Bug 2483: bind() called before connect()
Many of the occasions Squid was calling bind() are not required. This
reduces the bind() calls to only those which are actually needed.

Further optimization can be done in a future version to drop the paranoid
and slightly performance degrading safety checks for instances of Squid
binding ANYADDR without listener status, and attempting to bind NOADDR.