Author: Matthew Morgan <atcs.matthew@gmail.com> Add acl support to range_offset_limit.
It is being parsed using the same data types as reply_body_max_size,
which means the limit value gets run through parseBytesLine64, and the
acl is optional.

If people are using the old global style of range_offset_limit, they
shouldn't have to do anything to their configs.

getRangeOffsetLimit is now a member of HttpRequest, and it caches it's
value the same way reply_body_max_size does for faster repeat checking.