Bug 2633 fix: Ecap::HeaderRep::value(name) fails when there is no named header field
Calling Adaptation::Ecap::HeaderRep::value(const Name &name) with names
of header fields that do not exist leads to

    ICAP/AsyncJob.cc(218) dial: AsyncJob::noteStart threw exception:
    basic_string::_S_construct NULL not valid

I suspect this is a combination of
 - std::string constructor incapable of handling a nil char* pointer.
 - String::termedBuf() returning an nil pointer when the string is empty.

When there is no specified header field in the message, the value()
wrapper in Squid gets an empty String for the header value, which is
then used to create std::string, which fails or leads to failure.

I think it is wrong for termedBuf to return nil (because nil is not
0-terminated). I have not tried to fix that because we will have a new
String class soon.