Author:  Tsantilas Christos <chtsanti@users.sourceforge.net> , Alex Rousskov <rousskov@measurement-factory.com> Added %http::<bs and %icap::<bs logformat codes to HTTP and ICAP body sizes received from the next HTTP hop or the ICAP server.
Logging "received message body" is useful because Squid may receive a lot 
more or a lot less than it serves to the client or than the original resource
size, which may happen when handling Range requests and partial responses, 
when adapting bodies, and for other reasons.

For HTTP, we define "received message body" as message body bytes that 
Squid stores, merges, adapts, and/or forwards. In most cases, they are the 
same as body bytes sent by the server to Squid. However, the two bodies may 
differ for reasons such as errors (where the start of the body was not found),
HTTP transfer encodings (where Squid strips chunked encoding to find the 
message body), and generated FTP directory listings (that were received in 
a completely different format on a control connection).

For ICAP, the "received message body" is the Encapsulated sections, after 
the encapsulated HTTP body, if any, is dechunked.

This is a Measurement Factory project.