Make FTP able to retry after stupid FTP-protocol errors.
It only really matters if PUT/POST body data was sent to the server
Or if any reply headers/data was sent to the client.

For the relative long and error-prone state in between starting to FTP
connect and actually sending the data we can retry different FTP data
setup methods and forwarding paths forever.