Potential bug #3015 fix: assertion failed: comm.cc:143: "ccb->active()"
Fixed "ccb->active()" assertion related to the automatic write timeout.

We need to manually cancel writer's interest in select(2) before calling back
so that select() does not try to call the write handler when there is no
active callback anymore. Normally, select() cancels writer's interest when
calling the write handler, but in this case the call back is triggered not
from select() but from checkTimeouts().

From: 3p1-rock r9597