Author: Christian Wittmer <chris@computersalat.de> Add --with-swapdir=PATH to override default /var/cache/squid
For use on systems where /var/cache is not the required placement.

Adds the 'squid' part to the default path in accordance to the UNIX
standard filesystem layout definition. In squid-3 it is no longer used
for creating an implicit active cache. Only documentation and potential
core dumps use the variable set by this now.