Allow uri=value parameter when specifying adaptation service URIs.
When adaptation service URI contains a "=" character, Squid thinks you are
specifying a name=value option rather than a service URI. This leads to a
fatal configuration error. 

This change lets you specify the service URI using the uri=value
syntax where the value may contain "=" character(s). For example, the
following works after and only after this change:

  icap_service s1 reqmod_precache uri=icap://10.0.0.1:1344/adapt?bar=foo


TODO: Should the adaptation service parser be changed to treat the last word
on the icap_service line as a service URI, regardless of whether it contains
the "=" character?