Author: Fabian Hugelshofer <fh@open.ch> Allow persistent connections for Mozilla/3.0 User-Agents
This fixes NTLM and Negotiate authentication for these agents.

History:

In 1998 a hack was added to HttpMsg::persistent() that disables 
persistent connections for HTTP/1.0 User-Agents starting with 
"Mozilla/3." and "Netscape/3.".

According to the thread on squid-dev 
(http://www.eu.squid-cache.org/mail-archive/squid-dev/199805/0087.html), 
this was necessary to make some versions of Netscape browsers work that 
had a broken implementation of persistent connections. It was said that 
"NS 3.01 is ok. NS 3.02 is bad. NS 3.04 is good." Netscape 4 was ok, too.