Not all time-based options should accept milliseconds
After last changes, squid  accepts "millisecond" units for all time-based 
options, silently rounding millisecond-based values down if they are not 
supported. We should not allow millisecond units for options that do not 
support millisecond.
This patch add the allowMsec parameter to parseTimeLine function to allow/deny
the use of milliseconds.