Expand Makefile sources macros
Expand several macros used in earlier attempts to omtimize the Makefile
content. With the SourceLayout and modular changes underway these are
proving to be more of a problem than they are worth.

At some future time when the convenience libraries are settled it may be
worth revisiting some shared lists. But not yet.